# postman/openapi.json
{"openapi":"3.1.0","info":{"title":"KOINS Loan and Wallet Management API","description":"A fintech backend API for user onboarding, wallet funding, loan processing, repayments, and transaction tracking.","contact":{"name":"Oluwatosin Ajibade","email":"tosynajib@gmail.com"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"tags":[{"name":"Wallet","description":"Wallet balance, funding and transaction history APIs"},{"name":"Authentication","description":"Signup and login APIs"}],"paths":{"/api/v1/wallets/fund":{"post":{"tags":["Wallet"],"summary":"Create a wallet funding transaction","operationId":"fundWallet","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FundWalletRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseFundWalletResponse"}}}}}}},"/api/v1/auth/signup":{"post":{"tags":["Authentication"],"summary":"Create a new user account and wallet","operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseSignupResponse"}}}}}}},"/api/v1/auth/reset-password":{"post":{"tags":["Authentication"],"summary":"Reset password using OTP","operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseResetPasswordResponse"}}}}}}},"/api/v1/auth/resend-otp":{"post":{"tags":["Authentication"],"summary":"Resend password reset OTP","operationId":"resendOtp","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendOtpRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOtpResponse"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["Authentication"],"summary":"Login and receive JWT access token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseLoginResponse"}}}}}}},"/api/v1/auth/forgot-password":{"post":{"tags":["Authentication"],"summary":"Request password reset OTP","operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseOtpResponse"}}}}}}},"/api/v1/wallets/transactions":{"get":{"tags":["Wallet"],"summary":"Get authenticated user's wallet transaction history","operationId":"getMyWalletTransactions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseListTransactionResponse"}}}}}}},"/api/v1/wallets/me":{"get":{"tags":["Wallet"],"summary":"Get authenticated user's wallet","operationId":"getMyWallet","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseWalletResponse"}}}}}}},"/api/v1/wallets/balance":{"get":{"tags":["Wallet"],"summary":"Get authenticated user's wallet balance","operationId":"getMyWalletBalance","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseWalletBalanceResponse"}}}}}}}},"components":{"schemas":{"FundWalletRequest":{"type":"object","properties":{"amount":{"type":"number","minimum":100.00}},"required":["amount"]},"ApiResponseFundWalletResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/FundWalletResponse"},"timestamp":{"type":"string","format":"date-time"}}},"FundWalletResponse":{"type":"object","properties":{"reference":{"type":"string"},"amount":{"type":"number"},"currency":{"type":"string"},"status":{"type":"string"},"paymentAuthorizationUrl":{"type":"string"},"message":{"type":"string"}}},"SignupRequest":{"type":"object","properties":{"fullName":{"type":"string","maxLength":100,"minLength":2},"email":{"type":"string","format":"email","minLength":1},"phoneNumber":{"type":"string","minLength":1,"pattern":"^[0-9+]{10,15}$"},"password":{"type":"string","maxLength":2147483647,"minLength":8},"bvnOrNin":{"type":"string","maxLength":15,"minLength":10}},"required":["bvnOrNin","email","fullName","password","phoneNumber"]},"ApiResponseSignupResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/SignupResponse"},"timestamp":{"type":"string","format":"date-time"}}},"SignupResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"walletId":{"type":"string","format":"uuid"},"fullName":{"type":"string"},"email":{"type":"string"},"message":{"type":"string"}}},"ResetPasswordRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"otp":{"type":"string","maxLength":6,"minLength":6},"newPassword":{"type":"string","maxLength":2147483647,"minLength":8}},"required":["email","newPassword","otp"]},"ApiResponseResetPasswordResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ResetPasswordResponse"},"timestamp":{"type":"string","format":"date-time"}}},"ResetPasswordResponse":{"type":"object","properties":{"message":{"type":"string"}}},"ResendOtpRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1}},"required":["email"]},"ApiResponseOtpResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/OtpResponse"},"timestamp":{"type":"string","format":"date-time"}}},"OtpResponse":{"type":"object","properties":{"message":{"type":"string"},"expiresInMinutes":{"type":"integer","format":"int32"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password"]},"ApiResponseLoginResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginResponse"},"timestamp":{"type":"string","format":"date-time"}}},"LoginResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN"]},"accessToken":{"type":"string"},"tokenType":{"type":"string"},"expiresInMinutes":{"type":"integer","format":"int64"}}},"ForgotPasswordRequest":{"type":"object","properties":{"email":{"type":"string","format":"email","minLength":1}},"required":["email"]},"ApiResponseListTransactionResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}},"timestamp":{"type":"string","format":"date-time"}}},"TransactionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"walletId":{"type":"string","format":"uuid"},"type":{"type":"string","enum":["CREDIT","DEBIT","LOAN_DISBURSEMENT","REPAYMENT"]},"amount":{"type":"number"},"status":{"type":"string","enum":["PENDING","SUCCESSFUL","FAILED","REVERSED"]},"reference":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseWalletResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/WalletResponse"},"timestamp":{"type":"string","format":"date-time"}}},"WalletResponse":{"type":"object","properties":{"walletId":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"balance":{"type":"number"},"currency":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","SUSPENDED","CLOSED"]},"createdAt":{"type":"string","format":"date-time"}}},"ApiResponseWalletBalanceResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/WalletBalanceResponse"},"timestamp":{"type":"string","format":"date-time"}}},"WalletBalanceResponse":{"type":"object","properties":{"balance":{"type":"number"},"currency":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","name":"bearerAuth","scheme":"bearer","bearerFormat":"JWT"}}}}