# src/main/java/com/tosin/koins/auth/controller/AuthController.java
package com.tosin.koins.auth.controller;

import com.tosin.koins.auth.dto.*;
import com.tosin.koins.auth.facade.AuthFacade;
import com.tosin.koins.common.response.ApiResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

/**
 * Authentication controller.
 */
@RestController
@RequestMapping("/api/v1/auth")
@RequiredArgsConstructor
@Tag(name = "Authentication", description = "Signup and login APIs")
public class AuthController {

    private final AuthFacade authFacade;

    @PostMapping("/signup")
    @ResponseStatus(HttpStatus.CREATED)
    @Operation(summary = "Create a new user account and wallet")
    public ApiResponse<SignupResponse> signup(@Valid @RequestBody SignupRequest request) {
        return authFacade.signup(request);
    }

    @PostMapping("/login")
    @Operation(summary = "Login and receive JWT access token")
    public ApiResponse<LoginResponse> login(@Valid @RequestBody LoginRequest request) {
        return authFacade.login(request);
    }

    @PostMapping("/forgot-password")
    @Operation(summary = "Request password reset OTP")
    public ApiResponse<OtpResponse> forgotPassword(@Valid @RequestBody ForgotPasswordRequest request) {
        return authFacade.forgotPassword(request);
    }

    @PostMapping("/resend-otp")
    @Operation(summary = "Resend password reset OTP")
    public ApiResponse<OtpResponse> resendOtp(@Valid @RequestBody ResendOtpRequest request) {
        return authFacade.resendOtp(request);
    }

    @PostMapping("/reset-password")
    @Operation(summary = "Reset password using OTP")
    public ApiResponse<ResetPasswordResponse> resetPassword(@Valid @RequestBody ResetPasswordRequest request) {
        return authFacade.resetPassword(request);
    }

    @PostMapping("/logout")
    @Operation(summary = "Logout current user")
    public ApiResponse<Void> logout() {
        return authFacade.logout();
    }

}