# src/main/java/com/tosin/koins/auth/dto/ResendOtpRequest.java
package com.tosin.koins.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record ResendOtpRequest(@NotBlank(message = "Email is required")
                               @Email(message = "Email must be valid")
                               String email
) {
}